backend-facturacion/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
email-validator>=2.1.0
sqlmodel>=0.0.14
aiosqlite>=0.19.0
sqlalchemy[asyncio]>=2.0.0
cryptography>=41.0.0